requirements.txt
aioredis==2.0.0a1
asgiref==3.3.4
async-timeout==3.0.1
asyncstdlib==3.9.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==5.0.9
dhash==1.3
Django==4.0
future==0.18.2
geocoder==1.38.1
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.3
httptools==0.1.2
httpx==0.18.1
idna==2.10
orjson==3.5.2
Pillow==8.2.0
psycopg2-binary==2.8.6
pydantic==1.8.2
PyJWT==2.1.0
python-dotenv==0.17.0
pytz==2021.1
PyYAML==5.4.1
ratelim==0.1.6
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
